% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vs.r
\name{vs}
\alias{vs}
\title{Search for columns containing string}
\usage{
vs(chars = "", ds, ignore_case = TRUE)
}
\arguments{
\item{chars}{the character string to search for in the column names of
\code{ds}}

\item{ds}{object (i.e. data.frame) whose column names will be searched}

\item{ignore_case}{if \code{TRUE} then the case of \code{chars} and the
column names of \code{ds} are ignored, otherwise, it will only match
if the case matches}
}
\description{
\code{vs} returns a vector of all column names that contain the string
  \code{chars}
}
\details{
This function is meant to help you find the column name of interest when
  there are too many columns to visually search for the one you want
}
\examples{
data = data.frame(apple = 1:4,
                  orange = 5:8,
                  orangutan = 2:5,
                  applause = 3:6)

vs("oran", data)
}
